maul-py/requirements.txt
fastapi
uvicorn
redis
pgvector
psycopg2-binary
openai>=1.0.0
tiktoken
langchain>=0.1.16
langchain-community>=0.0.18
datasets
faker
python-dotenv
langchain-openai>=0.1.0
SQLAlchemy>=2.0
asyncpg>=0.27.0
httpx
python-multipart
aiofiles
